back db item table updated2

+++ database/migrations/2022_02_07_054109_create_items_table.php
@@ -16,15 +16,15 @@ class CreateItemsTable extends Migration
         Schema::create('items', function (Blueprint $table) {
             $table->id();
             $table->foreignId('category_id')->constrained()->cascadeOnDelete();
-            $table->string( 'short_name', 25 );
-            $table->string( 'name', 50);
+            $table->string('short_name', 25);
+            $table->string('name', 50);
             $table->integer('start_price');
             $table->string('short_desc', 50);
             $table->text('desc');
             $table->text('image_url');
             $table->text('animation_url')->nullable();
-            $table->string( 'type', 25 );
-            $table->tinyInteger('model_position');
+            $table->string('type', 25);
+            $table->float('model_position', 1, 1);
             $table->timestamps();
         });
     }
